# openapi.yaml
openapi: 3.1.0
info:
  title: Canopy Events API
  version: 1.0.0
  summary: Events among friends, for the web page and the iOS and Android apps.
  description: |
    The JSON API behind `events.canopysf.com`. The web pages use it with the
    `canopy_session` cookie; apps use it with `Authorization: Bearer <token>`,
    where the token is a `canopy_session` value. `docs/api.md` in the repo is
    the narrative guide: auth, the RSVP state machine, the visibility rules,
    pagination and errors, with worked examples.

    **Conventions.** JSON in and out. Times are ISO 8601 UTC strings
    (`2026-10-31T03:00:00.000Z`); input times need `Z` or an offset. Ids are
    strings. Errors are `{"error": "<a sentence>", "reason": "<snake_case>"}`
    with the right status: branch on `reason`, show `error` if you like.
    Lists take `?cursor=&limit=` and answer with `nextCursor` (null at the
    end). Every answer is `Cache-Control: no-store`.

    **Other people** are only ever `Person`: id, first and last name, short
    name and photo. Nobody's email, phone, Instagram, Venmo or Cash App is
    ever in an answer, except your own, from `GET /api/v1/me`.
  license:
    name: Private
    identifier: LicenseRef-Proprietary
servers:
  - url: https://events.canopysf.com
    description: Production
  - url: http://localhost:3001
    description: Local development
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /api/v1/events/{id}/rsvp:
    parameters:
      - $ref: '#/components/parameters/EventId'
    put:
      tags: [Guests]
      operationId: setRsvp
      summary: Answer, or change your answer
      description: |
        `going`, `maybe` or `not_going`. Anyone signed in with the link may
        answer, invited or not, unverified accounts included. Refused with
        409 for a cancelled event (`event_cancelled`), one that's over
        (`event_over`), or by a host (`host_cannot_rsvp`). With a
        `capacity`, a `going` that doesn't fit is saved as `waitlisted`;
        someone already going who asks for more plus-ones than there's
        room for gets 409 `no_room` and keeps their spot as it was.
        Changing to `maybe` or `not_going`, or fewer plus-ones, frees
        spots for the waitlist. `guests` (plus-
        ones) is up to the event's `guestsAllowed` (400 `too_many_guests`),
        0 if left out, and always 0 for `not_going`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RsvpInput' }
            examples:
              going:
                value: { status: going }
              withTwo:
                value: { status: going, guests: 2 }
      responses:
        '200':
          description: |
            The event after your answer; `viewer.rsvp` is it. `waitlisted`
            is true when you asked for `going` and the event was full (for
            you and your plus-ones), so you're on the waitlist instead:
            you'll be moved to `going` (and notified) when a spot that fits
            you frees up.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RsvpResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
    delete:
      tags: [Guests]
      operationId: withdrawRsvp
      summary: Take your answer back
      description: |
        If a host invited you, you're left invited with no answer;
        otherwise you're off the list. With no answer to take back it does
        nothing and still answers 200. The same 409s as answering. A spot
        you free goes to the waitlist.
      responses:
        '200':
          description: The event after; `viewer.rsvp` is `invited` or null.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EventEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/events/{id}/guests:
    parameters:
      - $ref: '#/components/parameters/EventId'
    get:
      tags: [Guests]
      operationId: listGuests
      summary: The guest list
      description: |
        Signed in only. `counts` always. Names (`guestsVisible: true`) for
        hosts always; for everyone else when the event's
        `guestListVisibility` is `everyone`, or once you've answered (any
        answer, `not_going` included) when it's `responded`. Otherwise
        `guestsVisible` is false and `guests` is empty. Non-hosts see
        answers only; hosts also see who's `invited` with no answer. In the
        order people got their status, oldest first.
      parameters:
        - name: status
          in: query
          description: Only this status. `invited` is for hosts (403 `hosts_only` otherwise).
          schema: { $ref: '#/components/schemas/RsvpStatus' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of the guest list.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GuestList' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/events/{id}/invites:
    parameters:
      - $ref: '#/components/parameters/EventId'
    post:
      tags: [Guests]
      operationId: invite
      summary: Invite people
      description: |
        Hosts only. By person id: the web page offers your friends
        (`GET /api/v1/me/friends`), but any Canopy account's id works.
        Each id ends up in `invited`, or in `skipped` with why:
        `already_on_list` (invited before, or already answered),
        `is_host`, or `not_found` (no such account). 1 to 100 ids a
        request; 300 invitations per host a day (429 `rate_limited`). 409
        for a cancelled event or one that's over.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InviteInput' }
            examples:
              two:
                value:
                  personIds: [6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44, 0b7e5a1f-9c2d-4e8b-8f3a-1d2c3b4a5e6f]
      responses:
        '200':
          description: Who was invited, and who wasn't and why.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InviteResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/EventNotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/events/{id}/invites/{personId}:
    parameters:
      - $ref: '#/components/parameters/EventId'
      - name: personId
        in: path
        required: true
        schema: { type: string }
    delete:
      tags: [Guests]
      operationId: uninvite
      summary: Take back an invitation
      description: |
        Hosts only, and only while it has no answer: 409
        `already_responded` once they've answered (their answer stays), 404
        `not_invited` if they weren't invited.
      responses:
        '200':
          description: Done.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ok' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/SignInRequired' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/Unavailable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        For apps. The token is a `canopy_session` value (43 characters,
        base64url), from the account service's native sign-in (still to be
        built). A request with a bearer header is signed in by it alone,
        never by a cookie, skips the Origin check, and never gets a
        `Set-Cookie` back. A malformed token means signed out.
    cookieAuth:
      type: apiKey
      in: cookie
      name: canopy_session
      description: |
        For the web pages. Set by `account.canopysf.com` for all of
        `canopysf.com`. Changes (POST, PUT, PATCH, DELETE) made with it
        must carry an `Origin` header for a Canopy page, or get 403
        `bad_origin`.
  parameters:
    EventId:
      name: id
      in: path
      required: true
      description: The event's id, 12 characters of base62 (the last part of its link).
      schema: { type: string, pattern: '^[0-9A-Za-z]{12}$' }
      example: 4fQ9xKpL2mZa
    Cursor:
      name: cursor
      in: query
      description: '`nextCursor` from the page before. Leave out (or empty) for the first page.'
      schema: { type: string }
    Limit:
      name: limit
      in: query
      description: How many to a page, 1 to 100.
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
  responses:
    BadRequest:
      description: |
        Something in the request isn't right: `bad_json`, `bad_title`,
        `bad_starts_at`, `bad_ends_at`, `ends_before_start`,
        `bad_time_zone`, `bad_guest_list_visibility`, `bad_description`,
        `bad_location_name`, `bad_location_address`, `bad_status`,
        `bad_guests`, `too_many_guests`, `bad_guests_allowed`,
        `bad_person_ids`, `bad_person_id`, `bad_text`, `bad_capacity`,
        `bad_image`, `bad_ids`, `bad_platform`, `bad_token`, `bad_cursor`,
        `bad_limit`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'an event has to end after it starts', reason: ends_before_start }
    SignInRequired:
      description: Not signed in (`sign_in_required`), with where to send someone.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: sign in first
            reason: sign_in_required
            signIn: https://account.canopysf.com/?return=https%3A%2F%2Fevents.canopysf.com%2Fe%2F4fQ9xKpL2mZa
            quickSignUp: https://account.canopysf.com/?quick=1&return=https%3A%2F%2Fevents.canopysf.com%2Fe%2F4fQ9xKpL2mZa
    Forbidden:
      description: |
        Not allowed: `hosts_only`, `creator_only` (co-hosts can't do this),
        `answer_first` (posting before saying you're going or maybe),
        `not_yours` (deleting someone else's post),
        `email_unverified` (with a `verify` link when it's about your own
        email; without one when it's about someone you tried to make a
        co-host), or `bad_origin` (a change made with the cookie from a
        page that isn't Canopy's).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            error: verify your email first
            reason: email_unverified
            verify: https://account.canopysf.com/profile?verify=1&return=https%3A%2F%2Fevents.canopysf.com%2F
    EventNotFound:
      description: No event at that id (`event_not_found`).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "there's no event at that link", reason: event_not_found }
    NotFound:
      description: '`event_not_found`, `not_invited`, `person_not_found`, `not_cohost` or `entry_not_found`.'
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: |
        Not in this event's state: `event_cancelled`, `event_over`,
        `host_cannot_rsvp`, `already_responded`, `is_creator`,
        `too_many_cohosts`, `no_room`.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: 'this event has been cancelled', reason: event_cancelled }
...
    RateLimited:
      description: Over a limit (`rate_limited`). Try again later.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: Canopy accounts couldn't be reached (`accounts_unreachable`). Try again in a minute.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error, reason]
      properties:
        error: { type: string, description: A sentence a person can read. }
        reason: { type: string, description: What happened, as a snake_case code to branch on. }
        signIn: { type: string, format: uri, description: 'With `sign_in_required`: where to sign in and come back.' }
        quickSignUp: { type: string, format: uri, description: 'With `sign_in_required`: the quick sign-up (name, email, passkey), for someone new to Canopy.' }
        verify: { type: string, format: uri, description: 'With `email_unverified`: where to prove the email and come back.' }
    Ok:
      type: object
      additionalProperties: false
      required: [ok]
      properties:
        ok: { const: true }
    Person:
      description: |
        Anyone other than you, and the only shape anyone else ever has. A
        deleted account (a former member) is `firstName: "Former member"`,
        `lastName: ""`, `shortName: "Former member"`, `photoUrl: null`.
      type: object
      additionalProperties: false
      required: [id, firstName, lastName, shortName, photoUrl]
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        shortName: { type: string, description: First name and last initial, e.g. "Ana L". }
        photoUrl:
          type: [string, 'null']
          format: uri
          description: |
            On account.canopysf.com. Served only with a Canopy session (the
            browser's cookie); null for no photo.
      example:
        id: 6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44
        firstName: Ana
        lastName: Lima
        shortName: Ana L
        photoUrl: https://account.canopysf.com/photo/6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44?v=1759870000000
...
    RsvpStatus:
      type: string
      enum: [invited, going, maybe, not_going, waitlisted]
      description: |
        `invited` is invited with no answer yet. `waitlisted` is a `going`
        that didn't fit the event's capacity; the server moves it to
        `going` when a spot that fits frees up. Nobody asks for it.
    Counts:
      description: |
        How many of each status. Always shown, to everyone. Hosts aren't
        counted. The top-level numbers are **people**; `guests` is the
        plus-ones those people bring, and `total` is the two together
        (people + plus-ones), for the answers that bring anyone.
      type: object
      additionalProperties: false
      required: [going, maybe, notGoing, invited, waitlisted, guests, total]
      properties:
        going: { type: integer, minimum: 0 }
        maybe: { type: integer, minimum: 0 }
        notGoing: { type: integer, minimum: 0 }
        invited: { type: integer, minimum: 0 }
        waitlisted: { type: integer, minimum: 0 }
        guests: { $ref: '#/components/schemas/GuestCounts' }
        total: { $ref: '#/components/schemas/GuestCounts' }
      example:
        going: 4
        maybe: 1
        notGoing: 1
        invited: 3
        waitlisted: 0
        guests: { going: 2, maybe: 0, waitlisted: 0 }
        total: { going: 6, maybe: 1, waitlisted: 0 }
    GuestCounts:
      type: object
      additionalProperties: false
      required: [going, maybe, waitlisted]
      properties:
        going: { type: integer, minimum: 0 }
        maybe: { type: integer, minimum: 0 }
        waitlisted: { type: integer, minimum: 0 }
    Host:
      type: object
      additionalProperties: false
      required: [person, role]
      properties:
        person: { $ref: '#/components/schemas/Person' }
        role:
          type: string
          enum: [creator, cohost]
          description: |
            The creator first, then co-hosts in the order they were added.
            Co-hosts can edit and invite; only the creator manages
            co-hosts, cancels and makes a new link.
    Rsvp:
      description: Your own place on an event.
      type: object
      additionalProperties: false
      required: [status, guests, guestsOverLimit, invited, respondedAt]
      properties:
        status: { $ref: '#/components/schemas/RsvpStatus' }
        guests: { type: integer, minimum: 0, description: Plus-ones you're bringing. }
        guestsOverLimit:
          type: boolean
          description: |
            True when the host has since lowered `guestsAllowed` below your
            `guests`. Your answer stands as it is; changing it has to fit.
        invited: { type: boolean, description: Whether a host invited you. }
        respondedAt: { type: [string, 'null'], format: date-time, description: When you last answered; null while only invited. }
    Viewer:
      description: You and this event. Null when signed out.
      type: object
      additionalProperties: false
      required: [role, rsvp, canEdit, canSeeGuestList, canPost]
      properties:
        role:
          type: [string, 'null']
          enum: [creator, cohost, null]
          description: Your part in hosting it, or null.
        rsvp:
          oneOf:
            - $ref: '#/components/schemas/Rsvp'
            - type: 'null'
        canEdit: { type: boolean }
        canSeeGuestList: { type: boolean, description: "Whether the guest list's names are shown to you. The wall follows the same rule." }
        canPost: { type: boolean, description: 'Whether you may post on the wall: hosts, and answers of going, maybe or waitlisted.' }
    FriendsGoing:
      description: |
        Which of your friends are going. `people` is the first 12, and
        empty when you can't see the guest list's names yet; `count` is
        always all of them.
      type: object
      additionalProperties: false
      required: [count, people]
      properties:
        count: { type: integer, minimum: 0 }
        people:
          type: array
          items: { $ref: '#/components/schemas/Person' }
    Event:
      type: object
      additionalProperties: false
      required: [id, url, title, description, startsAt, endsAt, timeZone, locationName, locationAddress,
                 locationAddressHidden, guestListVisibility, guestsAllowed, capacity, spotsLeft, coverImageUrl,
                 status, cancelledAt, createdAt, updatedAt,
                 hosts, counts, viewer]
      properties:
        id: { type: string, pattern: '^[0-9A-Za-z]{12}$' }
        url: { type: string, format: uri, description: The event's link, to share. }
        title: { type: string }
        description: { type: [string, 'null'] }
        startsAt: { type: string, format: date-time }
        endsAt: { type: [string, 'null'], format: date-time }
        timeZone: { type: string, description: 'IANA, e.g. America/Los_Angeles: show times in it.' }
        locationName: { type: [string, 'null'] }
        locationAddress: { type: [string, 'null'], description: Null when signed out (see locationAddressHidden). }
        locationAddressHidden: { type: boolean, description: True when there's an address you'll see once signed in. }
        guestListVisibility: { type: string, enum: [everyone, responded] }
        guestsAllowed: { type: integer, minimum: 0, maximum: 10, description: Plus-ones each answer may bring. }
        capacity:
          type: [integer, 'null']
          minimum: 1
          description: The most people going, plus-ones included; null for no cap.
        spotsLeft:
          type: [integer, 'null']
          minimum: 0
          description: '`capacity` minus `counts.total.going`, never below 0; null with no cap.'
        coverImageUrl:
          type: [string, 'null']
          format: uri
          description: |
            The cover image, a JPEG, or null. Public (no sign-in needed),
            so it works in link previews; it changes with every upload,
            so cache it by URL.
        status: { type: string, enum: [active, cancelled] }
        cancelledAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        hosts:
          type: array
          items: { $ref: '#/components/schemas/Host' }
        counts: { $ref: '#/components/schemas/Counts' }
        viewer:
          oneOf:
            - $ref: '#/components/schemas/Viewer'
            - type: 'null'
        friendsGoing:
          $ref: '#/components/schemas/FriendsGoing'
          description: Only on a single event (not in lists), and only when signed in.
      example:
        id: 4fQ9xKpL2mZa
        url: https://events.canopysf.com/e/4fQ9xKpL2mZa
        title: Rooftop dinner
        description: Bring a jacket.
        startsAt: '2026-11-01T02:30:00.000Z'
        endsAt: '2026-11-01T06:00:00.000Z'
        timeZone: America/Los_Angeles
        locationName: "Ana's place"
        locationAddress: 1 Market St, San Francisco
        locationAddressHidden: false
        guestListVisibility: responded
        guestsAllowed: 1
        capacity: 20
        spotsLeft: 14
        coverImageUrl: https://events.canopysf.com/covers/Qm7Zc2pR9xTa.jpg?v=1759870000000
        status: active
        cancelledAt: null
        createdAt: '2026-10-07T20:00:00.000Z'
        updatedAt: '2026-10-07T20:00:00.000Z'
        hosts:
          - person: { id: 6f1c2b9e-4d0a-4a53-9a51-2f7e0c1d8b44, firstName: Ana, lastName: Lima, shortName: Ana L, photoUrl: null }
            role: creator
        counts:
          going: 4
          maybe: 1
          notGoing: 1
          invited: 3
          waitlisted: 0
          guests: { going: 2, maybe: 0, waitlisted: 0 }
          total: { going: 6, maybe: 1, waitlisted: 0 }
        viewer:
          role: null
          rsvp: { status: going, guests: 1, guestsOverLimit: false, invited: true, respondedAt: '2026-10-08T17:12:00.000Z' }
          canEdit: false
          canSeeGuestList: true
          canPost: true
        friendsGoing:
          count: 1
          people:
            - { id: 0b7e5a1f-9c2d-4e8b-8f3a-1d2c3b4a5e6f, firstName: Ben, lastName: Okafor, shortName: Ben O, photoUrl: null }
    EventEnvelope:
      type: object
      additionalProperties: false
      required: [event]
      properties:
        event: { $ref: '#/components/schemas/Event' }
    RsvpResult:
      type: object
      additionalProperties: false
      required: [event, waitlisted]
      properties:
        event: { $ref: '#/components/schemas/Event' }
        waitlisted:
          type: boolean
          description: You asked for `going`, it was full, and you're on the waitlist instead.
...
    NextCursor:
      type: [string, 'null']
      description: Pass back as `?cursor=` for the next page; null at the end.
    Guest:
      type: object
      additionalProperties: false
      required: [person, status, guests, guestsOverLimit, respondedAt]
      properties:
        person: { $ref: '#/components/schemas/Person' }
        status: { $ref: '#/components/schemas/RsvpStatus' }
        guests: { type: integer, minimum: 0, description: Plus-ones they're bringing. }
        guestsOverLimit: { type: boolean, description: 'More than the event now allows (the host lowered `guestsAllowed` after they answered).' }
        respondedAt: { type: [string, 'null'], format: date-time }
    GuestList:
      type: object
      additionalProperties: false
      required: [guestsVisible, guests, counts, nextCursor]
      properties:
        guestsVisible:
          type: boolean
          description: False when you can't see names yet; then `guests` is empty and only `counts` says anything.
        guests:
          type: array
          items: { $ref: '#/components/schemas/Guest' }
        counts: { $ref: '#/components/schemas/Counts' }
        nextCursor: { $ref: '#/components/schemas/NextCursor' }
      example:
        guestsVisible: true
        guests:
          - person: { id: 0b7e5a1f-9c2d-4e8b-8f3a-1d2c3b4a5e6f, firstName: Ben, lastName: Okafor, shortName: Ben O, photoUrl: null }
            status: going
            guests: 1
            guestsOverLimit: false
            respondedAt: '2026-10-08T17:12:00.000Z'
        counts:
          going: 1
          maybe: 0
          notGoing: 0
          invited: 0
          waitlisted: 0
          guests: { going: 1, maybe: 0, waitlisted: 0 }
          total: { going: 2, maybe: 0, waitlisted: 0 }
        nextCursor: null
...
    RsvpInput:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [going, maybe, not_going] }
        guests: { type: integer, minimum: 0, maximum: 10, default: 0, description: "Plus-ones, up to the event's guestsAllowed. Ignored for not_going." }
...
    InviteInput:
      type: object
      required: [personIds]
      properties:
        personIds:
          type: array
          minItems: 1
          maxItems: 100
          items: { type: string, description: A Canopy person id. }
    InviteResult:
      type: object
      additionalProperties: false
      required: [invited, skipped]
      properties:
        invited:
          type: array
          items: { $ref: '#/components/schemas/Person' }
        skipped:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [personId, reason]
            properties:
              personId: { type: string }
              reason: { type: string, enum: [already_on_list, is_host, not_found] }